improved code structure and refactored code
