+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/components/AreaChart/utils.ts
@@ -1,0 +1,13 @@
+const formatNumber = (
+  rawValue: number | string,
+  options?: Intl.NumberFormatOptions | undefined,
+) => {
+  const value = Number(rawValue);
+  return Intl.NumberFormat('en', options).format(value);
+};
+
+export const toStandardAmount = (rawValue: number | string, decimals = 2) =>
+  formatNumber(rawValue, {
+    notation: 'standard',
+    maximumFractionDigits: decimals,
+  });
