--- superset-frontend/packages/superset-plugin-chart-immersatable/src/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useChartAccessorsAndScales';
