+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/types.ts
@@ -1,21 +1,3 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
 import {
   QueryFormData,
   supersetTheme,
@@ -29,12 +11,12 @@ import {
 } from '@superset-ui/core';
 import { ScaleOrdinal } from 'd3-scale';
 
-export interface SupersetPluginChartImmersatableStylesProps {
+export type SupersetPluginChartImmersatableStylesProps = {
   height: number;
   width: number;
   headerFontSize: keyof typeof supersetTheme.typography.sizes;
   boldText: boolean;
-}
+};
 
 interface SupersetPluginChartHelloWorldCustomizeProps {
   headerText: string;
@@ -58,7 +40,7 @@ export type TableColumnConfig = {
   truncateLongCells?: boolean;
   currencyFormat?: Currency;
 };
-export interface DataColumnMeta {
+export type DataColumnMeta = {
   key: string;
   label: string;
   dataType: GenericDataType;
@@ -71,7 +53,7 @@ export interface DataColumnMeta {
   isPercentMetric?: boolean;
   isNumeric?: boolean;
   config?: TableColumnConfig;
-}
+};
 
 export type SupersetPluginChartImmersatableProps =
   SupersetPluginChartImmersatableStylesProps &
@@ -86,7 +68,6 @@ export type SupersetPluginChartImmersatableProps =
       emitCrossFilters?: boolean;
       allowRearrangeColumns?: boolean;
       filters?: DataRecordFilters;
-      // add typing here for the props you pass in from transformProps.ts!
     };
 
 export type ChartGenericDataItem = Record<string, string | number>;
