+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/utils/processCustomData.ts
@@ -1,6 +1,6 @@
-interface CustomData {
+type CustomData = {
   [key: string]: any;
-}
+};
 
 const filterDataByDateInterval = (
   data: string,
